+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+import FlagsConsumer from "./src/FlagsConsumer";
+
 # Changelog
 All notable changes to this project are documented in this file.
+
+## [1.0.28] - 2019-06-27
+### Fix
+- Fix a problem when no children and no renderOn are provided to the FlagsConsumer.
 
 ## [1.0.27] - 2019-06-17
 ### Added